+++ src/ILP/default-ilp-client.ts
@@ -6,6 +6,7 @@ import GrpcIlpNetworkClientWeb from './grpc-ilp-network-client.web'
 import { AccountBalance } from './model/account-balance'
 import { PaymentResult } from './model/payment-result'
 import { PaymentRequest } from './model/payment-request'
+import IlpError from './ilp-error'
 
 class DefaultIlpClient implements IlpClientDecorator {
   /**
@@ -27,22 +28,21 @@ class DefaultIlpClient implements IlpClientDecorator {
   }
 
   /**
+   * This constructor is meant solely for testing purposes. Users should prefer
+   * DefaultIlpclient.defaultIlpClientWithEndpoint instead.
+   *
+   * @param networkClient A {@link IlpNetworkClient} which can make network calls to ILP infrastructure
+   */
+  public constructor(private readonly networkClient: IlpNetworkClient) {}
+
+  /**
    * Get the balance of the specified account on the connector.
    *
    * @param accountId The account ID to get the balance for.
    * @param accessToken Optional access token. If using node network client, accessToken must be supplied, otherwise
    *        it will be picked up from a cookie.
    * @return A Promise<AccountBalance> with balance information of the specified account
-   */
-  public constructor(private readonly networkClient: IlpNetworkClient) {}
-
-  /**
-   * Retrieve the balance for the given accountId.
-   *
-   * @param accountId The ILP accountId to retrieve a balance for.
-   * @param accessToken Optional access token. If using node network client, accessToken must be supplied, otherwise
-   *        it will be picked up from a cookie.
-   * @returns An AccountBalance with balance information of the specified account
+   * @throws an IlpError if the inputs were invalid or an error occurs during the call.
    */
   public async getBalance(
     accountId: string,
@@ -52,6 +52,9 @@ class DefaultIlpClient implements IlpClientDecorator {
     request.setAccountId(accountId)
     return this.networkClient
       .getBalance(request, accessToken)
+      .catch((error) => {
+        throw IlpError.from(error)
+      })
       .then((response) => AccountBalance.from(response))
   }
 
@@ -64,6 +67,7 @@ class DefaultIlpClient implements IlpClientDecorator {
    * @returns A promise which resolves to a `PaymentResult` of the original amount, the amount sent
    *        in the senders denomination, and the amount that was delivered to the recipient in their denomination, as
    *        well as if the payment was successful
+   * @throws an IlpError if the inputs were invalid or an error occurs during the call.
    */
   public async sendPayment(
     paymentRequest: PaymentRequest,
@@ -72,6 +76,9 @@ class DefaultIlpClient implements IlpClientDecorator {
     const request = paymentRequest.toProto()
     return this.networkClient
       .send(request, accessToken)
+      .catch((error) => {
+        throw IlpError.from(error)
+      })
       .then((response) => PaymentResult.from(response))
   }
 }