+++ src/XRP/xrp-client.ts
@@ -1,17 +1,20 @@
 import { Wallet } from 'xpring-common-js'
 import { BigInteger } from 'big-integer'
 import { XRPClientDecorator } from './xrp-client-decorator'
-import LegacyDefaultXRPClient from './legacy/legacy-default-xrp-client'
 import TransactionStatus from './transaction-status'
 import ReliableSubmissionXRPClient from './reliable-submission-xrp-client'
 import DefaultXRPClient from './default-xrp-client'
 import XRPClientInterface from './xrp-client-interface'
 import XRPTransaction from './model/xrp-transaction'
+import XRPLNetwork from '../Common/xrpl-network'
 
 /**
  * XRPClient is a client which interacts with the Xpring platform.
  */
 class XRPClient implements XRPClientInterface {
+  /** The XRPL Network of the node that this client is communicating with. */
+  public readonly network: XRPLNetwork
+
   private readonly decoratedClient: XRPClientDecorator
 
   /**
@@ -20,18 +23,16 @@ class XRPClient implements XRPClientInterface {
    * The XRPClient will use gRPC to communicate with the given endpoint.
    *
    * @param grpcURL The URL of the gRPC instance to connect to.
-   * @param useNewProtocolBuffers If `true`, then the new protocol buffer implementation from rippled will be used. Defaults to true.
+   * @param network The network this XRPClient is connecting to.
    * @param forceWeb If `true`, then we will use the gRPC-Web client even when on Node. Defaults to false. This is mainly for testing and in the future will be removed when we have browser testing.
    */
-  public constructor(
-    grpcURL: string,
-    useNewProtocolBuffers = true,
-    forceWeb = false,
-  ) {
-    const defaultXRPClient = useNewProtocolBuffers
-      ? DefaultXRPClient.defaultXRPClientWithEndpoint(grpcURL, forceWeb)
-      : LegacyDefaultXRPClient.defaultXRPClientWithEndpoint(grpcURL, forceWeb)
+  public constructor(grpcURL: string, network: XRPLNetwork, forceWeb = false) {
+    this.network = network
 
+    const defaultXRPClient = DefaultXRPClient.defaultXRPClientWithEndpoint(
+      grpcURL,
+      forceWeb,
+    )
     this.decoratedClient = new ReliableSubmissionXRPClient(defaultXRPClient)
   }