+++ src/index.ts
@@ -1,18 +1,32 @@
 export {
   AccountInfo,
   ClassicAddress,
+  PayIDUtils,
   SubmitSignedTransactionResponse,
   XRPAmount,
   Wallet,
   WalletGenerationResult,
   Utils,
 } from 'xpring-common-js'
-export { default as RippledFlags } from './XRP/rippled-flags'
-export { default as TransactionStatus } from './XRP/transaction-status'
-export { default as XRPClient } from './XRP/xrp-client'
 export { PaymentRequest, PaymentResult, AccountBalance, IlpClient } from './ILP'
 export { default as PayIDClient } from './PayID/pay-id-client'
 export { PayIDErrorType, default as PayIDError } from './PayID/pay-id-error'
 export { default as XRPLNetwork } from './Common/xrpl-network'
 export { default as ComplianceType } from './PayID/compliance-type'
 export { default as XpringClient } from './Xpring/xpring-client'
+export { default as XRPError, XRPErrorType } from './XRP/xrp-error'
+export { default as IlpError, IlpErrorType } from './ILP/ilp-error'
+
+export { RippledFlags, TransactionStatus, XRPClient } from './XRP'
+export {
+  XRPCurrencyAmount,
+  XRPCurrency,
+  XRPIssuedCurrency,
+  XRPMemo,
+  XRPPathElement,
+  XRPPath,
+  XRPPayment,
+  XRPSigner,
+  XRPTransactionType,
+  XRPTransaction,
+} from './XRP'