+++ test/PayID/pay-id-client-test.ts
@@ -5,7 +5,6 @@ import PayIDClient from '../../src/PayID/pay-id-client'
 import PayIDError, { PayIDErrorType } from '../../src/PayID/pay-id-error'
 import XRPLNetwork from '../../src/Common/xrpl-network'
 import SignatureWrapperInvoice from '../../src/PayID/Generated/model/SignatureWrapperInvoice'
-import SignatureWrapperCompliance from '../../src/PayID/Generated/model/SignatureWrapperCompliance'
 import ComplianceType from '../../src/PayID/compliance-type'
 
 // Parameters for getInvoice
@@ -43,7 +42,7 @@ describe('Pay ID Client', function(): void {
     const payIDClient = new PayIDClient(XRPLNetwork.Test)
 
     // WHEN an XRPAddress is requested for an invalid pay ID THEN an invalid payment pointer error is thrown.
-    payIDClient.xrpAddressForPayID(invalidPayID).catch((error) => {
+    payIDClient.addressForPayID(invalidPayID).catch((error) => {
       assert.equal(
         (error as PayIDError).errorType,
         PayIDErrorType.InvalidPaymentPointer,
@@ -73,7 +72,7 @@ describe('Pay ID Client', function(): void {
       })
 
     // WHEN an XRP address is requested.
-    const xrpAddress = await payIDClient.xrpAddressForPayID(payID)
+    const xrpAddress = await payIDClient.addressForPayID(payID)
 
     // THEN the address exists.
     // TODO(keefertaylor): Tighten up this condition when proper response parsing is implemented.
@@ -97,7 +96,7 @@ describe('Pay ID Client', function(): void {
       .reply(404, {})
 
     // WHEN an XRPAddress is requested.
-    payIDClient.xrpAddressForPayID(payID).catch((error) => {
+    payIDClient.addressForPayID(payID).catch((error) => {
       // THEN an unexpected response is thrown with the details of the error.
       assert.equal(
         (error as PayIDError).errorType,
@@ -128,7 +127,7 @@ describe('Pay ID Client', function(): void {
       .reply(serverErrorCode, serverError)
 
     // WHEN an XRPAddress is requested for a Pay ID.
-    payIDClient.xrpAddressForPayID(payID).catch((error) => {
+    payIDClient.addressForPayID(payID).catch((error) => {
       // THEN an unexpected response is thrown with the details of the error.
       assert.equal(
         (error as PayIDError).errorType,
@@ -159,7 +158,7 @@ describe('Pay ID Client', function(): void {
       .reply(serverErrorCode, serverError)
 
     // WHEN an XRPAddress is requested for a Pay ID.
-    payIDClient.xrpAddressForPayID(payID).catch((error) => {
+    payIDClient.addressForPayID(payID).catch((error) => {
       // THEN an unexpected response is thrown with the details of the error.
       assert.equal(
         (error as PayIDError).errorType,
@@ -193,7 +192,7 @@ describe('Pay ID Client', function(): void {
       })
 
     // WHEN an XRPAddress is requested for a Pay ID.
-    payIDClient.xrpAddressForPayID(payID).catch((error) => {
+    payIDClient.addressForPayID(payID).catch((error) => {
       // THEN an unexpected response is thrown.
       assert.equal(
         (error as PayIDError).errorType,
@@ -328,7 +327,7 @@ describe('Pay ID Client', function(): void {
     // THEN the invoice was the mocked response.
     assert.deepEqual(
       invoice,
-      SignatureWrapperCompliance.constructFromObject(mockResponse, null),
+      SignatureWrapperInvoice.constructFromObject(mockResponse, null),
     )
   })