+++ CHANGELOG.md
@@ -6,6 +6,28 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+# Added
+- `XRPTransaction` contains additional field `deliveredAmount` to represent the actual delivered amount of a transaction.
+
+## 4.2.0 - April 7, 2020
+
+This build contains fixes for generated code that may have prevented earlier versions of the 4.x library from building.
+
+## 4.1.0 - April 6, 2020
+
+# Added
+- `XRPTransaction` contains additional synthetic fields to represent the timestamp and hash of the transaction.
+
+## 4.0.0 - April 6, 2020
+
+## Changed
+- `XRPClient` requires a new parameter in it's constructor that identifies the network it is attached to.
+- `IlpClient` methods now throw `IlpError`s if something goes wrong during the call
+    (either client side or server side).
+    This is only breaking if users are handling special error cases, which were previously `grpc.ServiceError`s when
+    calling from node.js, and `grpc-web.Error`s in the browser.
+- `XRPClient` methods now throw `XRPError`s if something goes wrong during the call.
+
 
 ## 3.0.4 - March 24, 2020